+++ tree.h
@@ -7,16 +7,16 @@ enum Comparisson {SMALLER=0, EQUAL, BIGGER};
 
 class BTree {
 public:
-    BTree (int cName);
-    std::vector<Node*> search(int, Comparisson); //int queryValue, Comparisson queryType
-    Node* insert(int); //int queryValue
+    BTree (long long cName);
+    std::vector<Node*> search(long long, Comparisson); //int queryValue, Comparisson queryType
+    Node* insert(long long); //int queryValue
     void deleteNode(Node*, BTNode*);
     int size();
-    int getName();
+    long long getName();
 private:
     BTNode* root;
-    int columnName; // 0 means ID column
+    long long columnName; // 0 means ID column
     int sizee;
-    Node* add(int, BTNode*, BTNode*); //int queryValue, BTNode* curr
+    Node* add(long long, BTNode*, BTNode*); //int queryValue, BTNode* curr
     void deleteSingleNode(Node*, BTNode*); //Node* queryValue, BTNode* curr
 };
